remove openai caller

+++ process_test_articles.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import re
+import os
 import shutil
 from python_src.utils.database import DocDatabaseWhoosh
 
@@ -7,7 +8,8 @@ data = pd.read_csv('articles.csv', )
 data = data.astype(str)
 row_num, col_num = data.shape
 
-shutil.rmtree('./database')
+if os.path.exists('./database'):
+    shutil.rmtree('./database')
 doc_database = DocDatabaseWhoosh('./database')
 
 doc_database.add_batch((
